refactor: extract helpers from 5 oversized functions (Cycle 14)
- search_facets._matches_range_filter: 81L → 24L (7 check methods)
- dashboard/aggregator.aggregate: 78L → 28L (6 helper methods)
- content_scoring.score: 76L → 22L (2 helper methods)
- cli_recommend.recommend: 76L → 12L (4 module functions)
- content_summarizer.summarize: 72L → 11L (5 helper functions)

+++ personal_index/content_summarizer.py
@@ -86,7 +86,7 @@ def _score_sentence(sentence: str, word_freq: dict[str, int]) -> float:
     return score / len(words)
 
 
-def _score_and_select(sentences: list[str], word_freq: dict[str, float], max_sentences: int) -> list[str]:
+def _score_and_select(sentences: list[str], word_freq: dict[str, int], max_sentences: int) -> list[str]:
     scored: list[tuple[int, float, str]] = []
     for i, sentence in enumerate(sentences):
         score = _score_sentence(sentence, word_freq)
